refactor: rename ShakeButton to ShakeDemoButtons and update documentation and tutorial assets for the Screen Shake system
