+++ Samples~/ScreenShakeExample/Assets/Scripts/ShakeDemoButtons.cs
@@ -1,13 +1,16 @@
 using UnityEngine;
 using GameplayMechanicsUMFOSS.Systems;
 
-public class ShakeButton : MonoBehaviour
+namespace GameplayMechanicsUMFOSS.Samples.ScreenShake
 {
-    [SerializeField] public float magnitude ;
-    [SerializeField] public float duration ;
-
-    public void Trigger()
+    public class ShakeDemoButtons : MonoBehaviour
     {
-        ScreenShakeSystem_UMFOSS.Instance.TriggerShake(magnitude, duration);
+        [SerializeField] public float magnitude;
+        [SerializeField] public float duration;
+
+        public void Trigger()
+        {
+            ScreenShakeSystem_UMFOSS.Instance.TriggerShake(magnitude, duration);
+        }
     }
 }
