--- Samples~/ScreenShakeExample/ScriptExplainer.txt
@@ -1,101 +1,0 @@
-====================================================
-SCRIPT: ScreenShakeSystem_UMFOSS.cs
-AUTHOR: Your Name
-NAMESPACE: GameplayMechanicsUMFOSS.Systems
-==========================================
-
-PURPOSE:
-A singleton-based camera shake system that uses trauma and Perlin noise to produce smooth, procedural screen shake for both position and rotation. Includes configurable intensity, decay, and duration.
-
----
-
-Code    : `using UnityEngine;`
-Explain : Imports core Unity functionality needed for MonoBehaviour, Camera, Vector3, Quaternion, Mathf, and Time.
-
-Code    : `public static ScreenShakeSystem_UMFOSS Instance;`
-Explain : Implements a singleton so any script can call `TriggerShake()` without needing a direct reference.
-
-Code    : `[SerializeField] private float ShakeDecay, TraumaMultiplier, PositionMagnitude, RotationMagnitude;`
-Explain : Configurable parameters:
-
-* `ShakeDecay` → controls how quickly shake fades.
-* `TraumaMultiplier` → scales shake intensity dynamically.
-* `PositionMagnitude` & `RotationMagnitude` → maximum shake effect for camera.
-
-Code    : `void Awake()`
-Explain : Initializes singleton instance. Finds the main camera and stores its original position and rotation for smooth reset after shaking.
-
----
-
-### TriggerShake Function
-
-Code    : `public void TriggerShake(float magnitude, float duration)`
-Explain : Adds trauma to the camera shake system and sets the duration for shake effect.
-
-* `trauma = Mathf.Clamp01(trauma + magnitude);` → ensures trauma never exceeds 1.
-* `traumaTimeRemaining = duration;` → tracks how long shake should be applied.
-
----
-
-### Perlin Noise Functions for Smooth Shake
-
-Code    :
-```float GetFloat(float seed)
-{
-return (Mathf.PerlinNoise(seed, timeCounter) - 0.5f) * 2f;
-}```
-Explain : Generates a smooth pseudo-random value between -1 and 1 using Perlin noise. `timeCounter` makes the noise evolve over time, creating natural motion. `seed` allows different axes to have unique shake patterns.
-
-Code    :
-```Vector3 GetVec3()
-{
-return new Vector3(
-GetFloat(1),
-GetFloat(10),
-0f
-);
-}```
-Explain : Combines three Perlin noise values into a Vector3 for 2D shake (x and y). Ensures each axis moves differently for realistic shake.
-
----
-
-### Update Function – Applying Shake
-
-Code    :
-```if (traumaTimeRemaining > 0f && trauma > 0f)
-{
-traumaTimeRemaining -= Time.deltaTime;
-timeCounter += Time.deltaTime * Mathf.Pow(trauma, 0.5f) * TraumaMultiplier * 0.3f;
-float smoothTrauma = trauma * trauma;
-Vector3 offset = GetVec3() * PositionMagnitude * smoothTrauma;
-cam.transform.localPosition = originalPosition + offset;
-cam.transform.localRotation = Quaternion.Euler(offset * RotationMagnitude);
-trauma -= Time.deltaTime * ShakeDecay * (trauma + 0.3f);
-trauma = Mathf.Clamp01(trauma);
-}```
-
-1. `traumaTimeRemaining -= Time.deltaTime;` → decreases shake duration over time.
-2. `timeCounter += Time.deltaTime * Mathf.Pow(trauma, 0.5f) * TraumaMultiplier * 0.3f;` → advances Perlin noise smoothly; higher trauma increases shake speed.
-3. `float smoothTrauma = trauma * trauma;` → squares trauma for non-linear scaling.
-4. `Vector3 offset = GetVec3() * PositionMagnitude * smoothTrauma;` → calculates positional offset.
-5. `cam.transform.localPosition = originalPosition + offset;` → applies shake to camera position.
-6. `cam.transform.localRotation = Quaternion.Euler(offset * RotationMagnitude);` → rotates camera proportionally.
-7. `trauma -= Time.deltaTime * ShakeDecay * (trauma + 0.3f); trauma = Mathf.Clamp01(trauma);` → smooth decay of trauma.
-
----
-
-### Lerp Back to Original Position
-
-Code    :
-```cam.transform.localPosition = Vector3.Lerp(cam.transform.localPosition, originalPosition, Time.deltaTime * 5f);
-cam.transform.localRotation = Quaternion.Lerp(cam.transform.localRotation, originalRotation, Time.deltaTime * 5f);```
-Explain : Smoothly interpolates camera back to original position and rotation when shake ends. Prevents sudden snapping after trauma reaches zero.
-
----
-
-✅ Key Concepts:
-
-* Trauma → controls shake intensity and decay.
-* Perlin noise → smooth random offsets, avoids jittery movement.
-* Lerp → smoothly returns camera to original state after shake.
-* Squaring trauma → stronger shake at high trauma, smoother fade at low trauma.
